added index.js to utils and refactored code importing from utils
