+++ src/redux/Login.redux/Login.actions.js
@@ -1,6 +1,8 @@
 import axios from 'axios';
-import { getUserFromAuthToken } from '../../util/getUserWithAuthToken';
-import { sleep } from '../../util/sleep';
+import {
+  getUserFromAuthToken,
+  sleep,
+} from '../../util';
 import {
   PYTHON_API,
   setLoading,
